feat: Change sensors uid to address if have
